+++ tests/java/mcp/output/NumberSchema.java
@@ -6,13 +6,13 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class NumberSchema {
     @JsonProperty(value = "default")
-    public Integer default_;
+    public Long default_;
     @JsonProperty(value = "description")
     public String description;
     @JsonProperty(value = "maximum")
-    public Integer maximum;
+    public Long maximum;
     @JsonProperty(value = "minimum")
-    public Integer minimum;
+    public Long minimum;
     @JsonProperty(value = "title")
     public String title;
     @JsonProperty(value = "type", required = true)
