+++ tests/java/primitives/output/AllPrimitives.java
@@ -10,13 +10,13 @@ public class AllPrimitives {
     @JsonProperty(value = "floatField", required = true)
     public double floatField;
     @JsonProperty(value = "intField", required = true)
-    public int intField;
+    public long intField;
     @JsonProperty(value = "optionalBool")
     public Boolean optionalBool;
     @JsonProperty(value = "optionalFloat")
     public Double optionalFloat;
     @JsonProperty(value = "optionalInt")
-    public Integer optionalInt;
+    public Long optionalInt;
     @JsonProperty(value = "optionalString")
     public String optionalString;
     @JsonProperty(value = "stringField", required = true)
